requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
pydantic==2.8.2
requests==2.32.3
pillow==10.4.0
numpy==2.1.0
pydicom==2.4.4